requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
requests==2.31.0
pydantic==2.6.4
langchain==0.1.16
langchain-openai==0.1.3
python-dotenv==1.0.1
starlette==0.36.3
email-validator==2.0.0